+++ jni/test/boxed_test.dart
@@ -12,11 +12,8 @@ import 'test_util/test_util.dart';
 void main() {
   // Don't forget to initialize JNI.
   if (!Platform.isAndroid) {
-    try {
-      Jni.spawn(dylibDir: "build/jni_libs", jvmOptions: ["-Xmx128m"]);
-    } on JvmExistsException catch (_) {
-      // TODO(#51): Support destroying and reinstantiating JVM.
-    }
+    checkDylibIsUpToDate();
+    Jni.spawnIfNotExists(dylibDir: "build/jni_libs", jvmOptions: ["-Xmx128m"]);
   }
   run(testRunner: test);
 }
@@ -70,10 +67,8 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('JBoolean', () {
     using((arena) {
-      expect(JBoolean(true), JBoolean.trueValue);
-      expect(JBoolean(false), JBoolean.falseValue);
-      expect(JBoolean.falseValue.booleanValue(deleteOriginal: true), false);
-      expect(JBoolean.trueValue.booleanValue(deleteOriginal: true), true);
+      expect(JBoolean(false).booleanValue(deleteOriginal: true), false);
+      expect(JBoolean(true).booleanValue(deleteOriginal: true), true);
     });
   });
 }