+++ jni/test/jlist_test.dart
@@ -12,17 +12,14 @@ import 'test_util/test_util.dart';
 void main() {
   // Don't forget to initialize JNI.
   if (!Platform.isAndroid) {
-    try {
-      Jni.spawn(dylibDir: "build/jni_libs", jvmOptions: ["-Xmx128m"]);
-    } on JvmExistsException catch (_) {
-      // TODO(#51): Support destroying and reinstantiating JVM.
-    }
+    checkDylibIsUpToDate();
+    Jni.spawnIfNotExists(dylibDir: "build/jni_libs", jvmOptions: ["-Xmx128m"]);
   }
   run(testRunner: test);
 }
 
 void run({required TestRunnerCallback testRunner}) {
-  JList<JString> newJList(Arena arena) {
+  JList<JString> testDataList(Arena arena) {
     return [
       "1".toJString()..deletedIn(arena),
       "2".toJString()..deletedIn(arena),
@@ -33,13 +30,13 @@ void run({required TestRunnerCallback testRunner}) {
 
   testRunner('length get', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list.length, 3);
     });
   });
   testRunner('length set', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       list.length = 2;
       expect(list.length, 2);
       list.length = 3;
@@ -49,7 +46,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('[]', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list[0].toDartString(deleteOriginal: true), "1");
       expect(list[1].toDartString(deleteOriginal: true), "2");
       expect(list[2].toDartString(deleteOriginal: true), "3");
@@ -57,7 +54,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('[]=', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list[0].toDartString(deleteOriginal: true), "1");
       list[0] = "2".toJString()..deletedIn(arena);
       expect(list[0].toDartString(deleteOriginal: true), "2");
@@ -65,7 +62,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('add', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       list.add("4".toJString()..deletedIn(arena));
       expect(list.length, 4);
       expect(list[3].toDartString(deleteOriginal: true), "4");
@@ -73,8 +70,8 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('addAll', () {
     using((arena) {
-      final list = newJList(arena);
-      final toAppend = newJList(arena);
+      final list = testDataList(arena);
+      final toAppend = testDataList(arena);
       list.addAll(toAppend);
       expect(list.length, 6);
       list.addAll(["4".toJString()..deletedIn(arena)]);
@@ -83,7 +80,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('clear, isEmpty, isNotEmpty', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list.isNotEmpty, true);
       expect(list.isEmpty, false);
       list.clear();
@@ -93,7 +90,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('contains', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       // ignore: iterable_contains_unrelated_type
       expect(list.contains("1"), false);
       expect(list.contains("1".toJString()..deletedIn(arena)), true);
@@ -102,7 +99,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('getRange', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       // ignore: iterable_contains_unrelated_type
       final range = list.getRange(1, 2)..deletedIn(arena);
       expect(range.length, 1);
@@ -111,7 +108,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('indexOf', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list.indexOf(1), -1);
       expect(list.indexOf("1".toJString()..toDartString()), 0);
       expect(list.indexOf("2".toJString()..toDartString()), 1);
@@ -121,7 +118,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('insert', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       list.insert(1, "0".toJString()..deletedIn(arena));
       expect(list.length, 4);
       expect(list[1].toDartString(deleteOriginal: true), "0");
@@ -129,8 +126,8 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('insertAll', () {
     using((arena) {
-      final list = newJList(arena);
-      final toInsert = newJList(arena);
+      final list = testDataList(arena);
+      final toInsert = testDataList(arena);
       list.insertAll(1, toInsert);
       expect(list[1].toDartString(deleteOriginal: true), "1");
       expect(list.length, 6);
@@ -141,7 +138,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('iterator', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       final it = list.iterator;
       expect(it.moveNext(), true);
       expect(it.current.toDartString(deleteOriginal: true), "1");
@@ -154,7 +151,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('remove', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list.remove("3".toJString()..deletedIn(arena)), true);
       expect(list.length, 2);
       expect(list.remove("4".toJString()..deletedIn(arena)), false);
@@ -164,22 +161,22 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('removeAt', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       expect(list.removeAt(0).toDartString(deleteOriginal: true), "1");
       expect(list.removeAt(1).toDartString(deleteOriginal: true), "3");
     });
   });
   testRunner('removeRange', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       list.removeRange(0, 2);
       expect(list.single.toDartString(deleteOriginal: true), "3");
     });
   });
   testRunner('==, hashCode', () {
     using((arena) {
-      final a = newJList(arena);
-      final b = newJList(arena);
+      final a = testDataList(arena);
+      final b = testDataList(arena);
       expect(a.hashCode, b.hashCode);
       expect(a, b);
       b.add("4".toJString()..deletedIn(arena));
@@ -189,7 +186,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner('toSet', () {
     using((arena) {
-      final list = newJList(arena);
+      final list = testDataList(arena);
       final set = list.toSet()..deletedIn(arena);
       expect(set.length, 3);
     });